+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `aes.deepDecrypt` and `aes.deepEncrypt` now use generics for better return typing
+- Configured terser to remove error messages for minified builds
+
+### Removed
+- Parsing of 'true' or 'false' as booleans in `aes.decrypt`, applications should handle this manually on a case by case basis
+
 ## [0.7.1]
 ### Fixed
 - Base64 decoding now ignores newlines, which fixes decoding long base64 strings produced many standard implementations of base64 (including MySQL's `TO_BASE64()`)