+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+[0.7.3]
 ### Added
 - `aes.deepDecrypt` and `aes.deepEncrypt` now use generics for better return typing
 - Configured terser to remove error messages for minified builds
@@ -14,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Parsing of 'true' or 'false' as booleans in `aes.decrypt`, applications should handle this manually on a case by case basis
+
+### Fixed
+- ESlint errors caused by TypeScript-incompatible rules
 
 ## [0.7.1]
 ### Fixed